+++ Ex1.java
@@ -1,10 +1,5 @@
 import java.io.FileNotFoundException;
 import java.io.IOException;
-import java.lang.reflect.Array;
-import java.util.Arrays;
-import java.util.HashSet;
-import java.util.LinkedList;
-import java.util.Queue;
 
 /* =========Ex1============
  *
@@ -54,7 +49,7 @@ public class Ex1
         System.out.println();
     }
     //Manhattan function as my heuristic function (useful for 8-tile puzzle)
-    public static int heuristic(Node n)
+    public static double heuristic(Node n)
     {
         //if the node is G - return 0
         if(n.getTag() == 'G')
@@ -66,7 +61,7 @@ public class Ex1
         //update distance to be according to the distance between start to target
         //int distance = Math.abs(i - row_goal) + Math.abs(j - column_goal); // Manhattan - Option A
         //int distance = Math.max(Math.abs(i - row_goal), Math.abs(j - column_goal)); //Diagonal distance - Option B
-        int distance = (int)Math.sqrt(Math.pow(i-row_goal,2)+Math.pow(j-column_goal,2)); //Euclid - Option C
+        double distance = Math.sqrt(Math.pow(i-row_goal,2)+Math.pow(j-column_goal,2)); //Euclid - Option C
         return distance;
     }
     private static void introToUser()
